// faberHouse/src/entities/project/model/types.ts
export interface IProject {
  id?: number;
  title: string;
  series: string;
  technology: string;
  size: string;
  additionally: string;
  roomCount: number;
  square: number;
  price: number;
  images: string[];
  videos: string[];
  files: string[];
  foundation: string;
  roomHeight: string;
  constructionTechnology: string;
  additionallyDescription: string;
}